ba: General updates
-Deletes `/build` from `.gitignore`.
-Moves `NavDrawer.jsx` into common folder.
-Deletes unused/old boilerplate code.
