--- app/assets/apps/bidsketch/src/components/Navigation/common/Navbar.css
@@ -1,3 +1,0 @@
-.aaws-logo {
-    position: absolute;
-}
