+++ app/assets/apps/bidsketch/src/components/common/NavDrawer/css/Navbar.css
@@ -1,0 +1,7 @@
+.bidsketch-logo {
+    position: absolute;
+}
+
+.nav-list-item {
+    text-decoration-line: none;
+}
