+++ app/assets/apps/bidsketch/src/components/common/NavDrawer/images/bidsketch-logo.png
