--- app/assets/apps/bidsketch/src/components/views/BidSketchFlow/BidSketchFlow.jsx
@@ -1,11 +1,0 @@
-import React, { Component } from "react";
-
-export default class BidSketchFlow extends Component {
-    render() {
-        return (
-            <div>
-                BidSketchFlow
-            </div>
-        );
-    }
-}
