+++ app/assets/apps/bidsketch/src/components/views/NotFound/css/NotFound.css
@@ -1,0 +1,3 @@
+.not-found-container {
+    text-align: center;
+}
